% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_permutations.R
\name{get_permutations}
\alias{get_permutations}
\title{Get all the unique combinations of questions from a list of survey questions}
\usage{
get_permutations(question_list, numeric_indicators = TRUE)
}
\arguments{
\item{...}{A list of questions or survey modules}

\item{type}{The type of question to create ("Open", etc.)}
}
\value{
A survey question
}
\description{
Get all the unique combinations of questions from a list of survey questions
}

