% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert something between questions}
\usage{
insert(image_path = NULL, space_reduction = NULL, page_break = FALSE,
  line = FALSE, custom = NULL, concat = FALSE, ...)
}
\arguments{
\item{space_reduction}{Put in a spatial measure to reduce space between questions}

\item{page_break}{Insert a page break}

\item{line}{Insert a line}

\item{custom}{Insert whatever text or functions you like}
}
\value{
An argument or text between questions
}
\description{
Insert something between questions
}

