% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_image}
\alias{insert_image}
\title{Insert an image}
\usage{
insert_image(image_path, caption = NULL, float_environment = "h!",
  image_width = "\\\\textwidth", centering = TRUE, caption_above = TRUE,
  concat = FALSE)
}
\arguments{
\item{image_path}{The path to the image on the computer (including name and extension of image, eg. image.pdf)}

\item{caption}{Caption to put on image (optional)}

\item{float_environment}{Environment to float image in (defaults to "h!")}

\item{image_width}{Width of image (defaults to "\\textwidth")}

\item{centering}{Should image be centered?}
}
\value{
An image in the survey
}
\description{
Insert an image
}

