% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_question.R
\name{make_horizontal_question}
\alias{make_horizontal_question}
\title{Create a horizontal multiple choice question}
\usage{
make_horizontal_question(question_prompt = NULL, comment = NULL,
  upper_labels = NULL, lower_labels = NULL, concat = F)
}
\arguments{
\item{question_prompt}{The main text of the question.}

\item{concat}{Logical argument: should the text be concatenated and printed using cat(), i.e. for sinking directly to .tex files.}
}
\value{
An open survey question.
}
\description{
Create a horizontal multiple choice question
}
\examples{
make_open_question("How are you?")
}

