% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_module.R
\name{make_module}
\alias{make_module}
\title{Make question module}
\usage{
make_module(..., title = NULL)
}
\arguments{
\item{...}{A list of questions}

\item{title}{The title of the module}
}
\value{
A question module
}
\description{
Make question module
}

