% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_question.R
\name{make_prompt}
\alias{make_prompt}
\title{Create a prompt}
\usage{
make_prompt(prompt = NULL, concat = F)
}
\arguments{
\item{prompt}{The non-question prompt or LaTeX formatting}
}
\value{
A prompt.
}
\description{
Create a prompt
}

