% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_question.R
\name{make_question}
\alias{make_question}
\title{Create a question.}
\usage{
make_question(question_prompt, type = "open", answer_categories = NULL,
  comment = NULL)
}
\arguments{
\item{question_prompt}{The main text of the question.}

\item{type}{The type of question to create ("Open", etc.)}
}
\value{
A survey question
}
\description{
Create a question.
}

