% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_survey.R
\name{make_survey}
\alias{make_survey}
\title{Output a survey into .tex or .pdf format}
\usage{
make_survey(..., output_file_path = NULL, N_surveys = NULL,
  survey_title = NULL, ID = NULL, compile_PDF = TRUE, store_tex = FALSE,
  page_numbers = TRUE)
}
\arguments{
\item{...}{A list of questions or survey modules}

\item{type}{The type of question to create ("Open", etc.)}
}
\value{
A survey question
}
\description{
Output a survey into .tex or .pdf format
}

