% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_question.R
\name{make_vertical_question}
\alias{make_vertical_question}
\title{Create a vertical multiple choice question}
\usage{
make_vertical_question(question_prompt = NULL, comment = NULL,
  labels = NULL, concat = F, image_path = NULL, ...)
}
\arguments{
\item{question_prompt}{The main text of the question.}

\item{concat}{Logical argument: should the text be concatenated and printed using cat(), i.e. for sinking directly to .tex files.}
}
\value{
An open survey question.
}
\description{
Create a vertical multiple choice question
}
\examples{
make_open_question("How are you?")
}

