% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert numbers to english words}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{A number}
}
\value{
A number spelled out in English words
}
\description{
Convert numbers to english words
}

