% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomize_survey.R
\name{randomize_survey}
\alias{randomize_survey}
\title{Randomize survey}
\usage{
randomize_survey(..., output_file_path = NULL, question_combinations = NULL,
  random_method = "simple", combination_probs = NULL,
  combination_sizes = NULL, N_surveys = NULL, survey_title = NULL,
  ID = NULL, compile_PDF = TRUE, store_tex = FALSE, page_numbers = TRUE)
}
\arguments{
\item{...}{A list of all questions or modules, with every possible permutation of every randomized question. Careful attention should be paid to the indexing of this list, as it will be used by the 'question_combinations' object to create full surveys.}

\item{question_combinations}{A list or matrix of vectors, whose elements index the question/module list, stating the unique combinations questions to be randomized. Each element of the list or column of the matrix should be a vector of integers, where each integer indexes a unique element in the list of questions or modules. Thus, each vector of integers represents a complete survey.}

\item{random_method}{Define the randomization method. If "complete", then each unique vector in question_combinations is repeated exactly combination_sizes times. If "simple", then each unique vector is generated as a survey with equal probabilities or combination_probs, until the survey is of size N_surveys or sum(combination_sizes). If "asis", randomize_survey simply applies through the question_combinations argument. This method is useful for more complex randomization schemes, where a pre-randomized question_combinations argument is provided.}

\item{combination_probs}{If "simple" is provided to random_method, column or list element i of question_combinations is chosen with probability combination_probs[i].}

\item{combination_sizes}{If "simple" is provided to random_method, column or list element i of question_combinations is assigned to combination_sizes[i] units in the sample.}

\item{N_surveys}{The number of surveys to generate. Should be left blank if combination_sizes is provided.}

\item{ID}{An optional ID variable to replace the automatically generated one. Should be as long as the number of surveys (i.e. N_surveys or sum(combination_sizes)).}
}
\value{
Outputs all of the uniquely-identified pdfs or .tex files for use in a survey experiment.
}
\description{
Randomize survey
}

