% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_question.R
\name{make_open_question}
\alias{make_open_question}
\title{Create an open question.}
\usage{
make_open_question(question_prompt, comment = NULL, concat = F,
  N_lines = NULL, N_dot_lines = NULL, image_path = NULL, ...)
}
\arguments{
\item{question_prompt}{The main text of the question.}

\item{concat}{Logical argument: should the text be concatenated and printed using cat(), i.e. for sinking directly to .tex files.}

\item{N_lines}{How many lines to print for the response}

\item{N_dot_lines}{How many dotted lines to print for the response}
}
\value{
An open survey question.
}
\description{
Create an open question.
}
\examples{
make_open_question("How are you?")
}

